pyproject.toml
[tool.poetry]
name = "Floccose"
version = "0.1.0"
description = ""
authors = ["UstmanAbdurahman <72262190+UstmanAbdurahman@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.2"
requests = "^2.25.1"
sentry-sdk = "^1.1.0"
firebase-admin = "^5.0.0"

[tool.taskipy.tasks]
start = "python -m bot"
lint = "pre-commit run --all-files"
precommit = "pre-commit install"

[tool.poetry.dev-dependencies]
flake8 = "~=3.8"
flake8-annotations = "~=2.3"
flake8-bugbear = "~=20.1"
flake8-docstrings = "~=1.5"
flake8-import-order = "~=0.18"
flake8-string-format = "~=0.3"
flake8-tidy-imports = "~=4.1"
flake8-todo = "~=0.7"
pep8-naming = "~=0.11"
pre-commit = "~=2.1"
python-dotenv = "^0.15.0"
taskipy = "^1.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"